+++ backend/src/main/java/edu/jmi/openatom/server/openatomsystem/dto/RequestPointRuleSettingsDTO.java
@@ -5,6 +5,6 @@ import lombok.Data;
 /** 积分规则配置请求 */
 @Data
 public class RequestPointRuleSettingsDTO {
-  private Integer dailyLoginPoints;
-  private Integer blogPublishPoints;
+  private Long dailyLoginPoints;
+  private Long blogPublishPoints;
 }
