+++ backend/src/main/java/edu/jmi/openatom/server/openatomsystem/service/impl/CheckInServiceImpl.java
@@ -519,7 +519,7 @@ public class CheckInServiceImpl implements CheckInService {
     return StpUtil.isLogin() ? StpUtil.getLoginIdAsInt() : null;
   }
 
-  private int safePoints(Integer value) {
-    return value == null ? 0 : Math.max(0, value);
+  private long safePoints(Long value) {
+    return value == null ? 0L : Math.max(0L, value);
   }
 }
